requirements.txt
Django==3.2
psycopg2==2.9.5
gunicorn==20.1.0
python-telegram-bot==12.8
python-dotenv==0.21.1
Pillow==9.4.0
channels==3.0.4
channels-redis==3.3.1